Perfiles funcionando correctamente

+++ src/main/java/co/edu/inmobiliaria/pau/iva/backendpauiva/Infraestructura/ClientesRepository.java
@@ -3,6 +3,6 @@ package co.edu.inmobiliaria.pau.iva.backendpauiva.Infraestructura;
 import co.edu.inmobiliaria.pau.iva.backendpauiva.Dominio.Clientes;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClientesRepository extends JpaRepository<Clientes, Long> {
+public interface ClientesRepository extends JpaRepository<Clientes, Integer> {
     Clientes findByCorreo(String correo);
 }
